WIP - User is prompted for reason for deletion
