+++ src/lib/urls.ts
@@ -26,7 +26,7 @@ export enum PageURLs {
   emergencyContact = "/register-a-beacon/emergency-contact",
   checkYourAnswers = "/register-a-beacon/check-your-answers",
   applicationComplete = "/register-a-beacon/application-complete",
-  deleteRegistration = "/manage-my-beacons/delete-a-registration",
+  deleteRegistration = "/manage-my-registrations/delete",
   serverError = "/500",
 }
